Add CheckBox Recognition
added some more classes and changed how the JSON is formatted. Instead
of sending the key, it now sends a TEMPLATE string array with all of the
names, and another int array with simple ones and zeroes. This may
change later to true or false, depending on what the end software wants.
This is all just to be more compatable with spreadsheets.